+++ orderComponent/src/main/java/com/vcs/bmp/microservices/order/repository/CustomOrderRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface CustomOrderRepository {
 
-    Integer countOrdersByCustomerIdAndAccountIdAndOrderItemsSkuIn(@NonNull String customerId,
+    Long countOrdersByCustomerIdAndAccountIdAndOrderItemsSkuIn(@NonNull String customerId,
                                                              @NonNull  String accountId ,
                                                              @NonNull List<String> itemSkus,
                                                              @Nullable ContextInfo contextInfo);
